move non-scipy dependent checks to test_core
