+++ plotly/tests/test_optional/test_opt_tracefactory.py
@@ -8,41 +8,8 @@ from nose.tools import raises
 
 import numpy as np
 
-import scipy
-import scipy.stats
-
 
 class TestDistplot(TestCase):
-
-    def test_wrong_curve_type(self):
-
-        # check: PlotlyError (and specific message) is raised if curve_type is
-        # not 'kde' or 'normal'
-
-        kwargs = {'hist_data': [[1, 2, 3]], 'group_labels': ['group'],
-                  'curve_type': 'curve'}
-        self.assertRaisesRegexp(PlotlyError, "curve_type must be defined as "
-                                             "'kde' or 'normal'",
-                                tls.FigureFactory.create_distplot, **kwargs)
-
-    def test_wrong_histdata_format(self):
-
-        # check: PlotlyError if hist_data is not a list of lists or list of
-        # np.ndarrays (if hist_data is entered as just a list the function
-        # will fail)
-
-        kwargs = {'hist_data': [1, 2, 3], 'group_labels': ['group']}
-        self.assertRaises(PlotlyError, tls.FigureFactory.create_distplot,
-                          **kwargs)
-
-    def test_unequal_data_label_length(self):
-        kwargs = {'hist_data': [[1, 2]], 'group_labels': ['group', 'group2']}
-        self.assertRaises(PlotlyError, tls.FigureFactory.create_distplot,
-                          **kwargs)
-
-        kwargs = {'hist_data': [[1, 2], [1, 2, 3]], 'group_labels': ['group']}
-        self.assertRaises(PlotlyError, tls.FigureFactory.create_distplot,
-                          **kwargs)
 
     def test_simple_distplot(self):
